+++ frontend/src/components/subComponents/LineGraph.js
