+++ frontend/src/components/subComponents/RangeTimeSeriesSelector.js
